implement Haptic Socket and plug channels
these can be animated. Animating a socket channel will alter its light range. 

I wanted to make it also possible to animate the ring/hole property, but it got a bit messy in the code, so i reverted and let it at this. I hope i can get that also animated later down the line somehow

Signed-off-by: Tayou <git@tayou.org>