Bugfix: Fix DirectTreeOptimizer for clips with partial keyframes
When animating for example the rotation of an
object, it's possible to skip the "end" keyframe on some of
the properties and Unity uses the value of the first keyframe
for the rest of the curve.

For the Direct Tree Optimizer, the end clip
is calculated from the curves, however, it does
currently not take into account those clips
that only have a "start" keyframe and no "end" keyframe.

This change looks at the curves and if there's only one keyframe,
it copies the value of the "start" clip to the "end" clip.