fix: disable NdmfFirstMenuItem for NDMF 1.3.0-rc.* or newer
With NDMF 1.3.0, I intend to adjust the build hook order so that the NdmfFirstMenuItem
hack is not necessary. Further, executing all of NDMF in one go can cause compatibility
issues with optimization plugins, as noted in #216.

Once #219 goes in, and with NDMF 1.3.0, it is now possible to have good compatibility
between VRCF and NDMF/MA by using Av3 Emulator with preprocess hooks enabled. As such,
this hack is no longer necessary and should be disabled. To ease the transition, this
PR conditionally disables this hack when NDMF 1.3.0+ is installed, to allow for time
for folks to update. Eventually, this class should be deleted entirely.